requirements.txt
pillow==10.3.0
Django==5.0.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
drf-spectacular==0.27.2
